server: move createSandboxContainer and related functions to container_create.go
This allows most of the container create logic to be shared between
Linux and FreeBSD.
